+++ frontend/src/index.js
@@ -71,7 +71,6 @@ window.Vue.mixin({
         },
 
         /**
-         *
          * @param {int} characterId
          * @param {function} [callback]
          */
@@ -96,6 +95,27 @@ window.Vue.mixin({
                 } else {
                     vm.message('Update done.', 'success');
                 }
+                if (typeof callback === typeof Function) {
+                    callback();
+                }
+            });
+        },
+
+        /**
+         * @param {int} characterId
+         * @param {int} [admin]
+         * @param {function} [callback]
+         */
+        deleteCharacter(characterId, admin, callback) {
+            const vm = this;
+            vm.loading(true);
+            new this.swagger.PlayerApi().deleteCharacter(characterId, { admin: admin || 0 }, function(error) {
+                vm.loading(false);
+                if (error) { // 403 usually
+                    vm.message('Deletion denied.', 'error');
+                    return;
+                }
+                vm.message('Deleted character.', 'success');
                 if (typeof callback === typeof Function) {
                     callback();
                 }